requirements_frozen.txt
# ============================================
# OffGallery - Dipendenze CONGELATE
# Versioni esatte testate e funzionanti
# Data freeze: Febbraio 2026
# ============================================

# === PYTORCH (CUDA 11.8) ===
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.7.1+cu118
torchvision==0.22.1+cu118

# === TRANSFORMERS / HUGGINGFACE ===
transformers==4.57.3
huggingface-hub==0.36.0
safetensors==0.7.0
tokenizers==0.22.1
timm==1.0.22

# === BIOCLIP ===
pybioclip==2.1.1
open-clip-torch==3.2.0

# === IMAGE PROCESSING ===
pillow==12.0.0
rawpy==0.25.1
opencv-contrib-python==4.12.0.88
numpy==2.2.6
imageio==2.37.2
scikit-image==0.25.2

# === TRANSLATION ===
argostranslate==1.10.0
ctranslate2==4.6.1
sentencepiece==0.2.1

# === UI ===
PyQt6==6.10.0
PyQt6-Qt6==6.10.1
PyQt6-sip==13.10.2

# === UTILITIES ===
pyyaml==6.0.3
requests==2.32.5
tqdm==4.67.1
ftfy==6.3.1
regex==2025.11.3
pandas==2.3.3
scipy==1.16.3
scikit-learn==1.7.2

# === SUPPORTO ===
filelock==3.19.1
fsspec==2025.9.0
packaging==25.0
typing-extensions==4.15.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0

# ============================================
# NOTE:
# - Queste versioni sono testate insieme
# - NON aggiornare senza testare l'app
# - Per aggiornare: crea un nuovo ambiente di test
# ============================================
